Pioneer

Documentation for Pioneer.

Pioneer.FragmentIndex

Pioneer.FragmentMatch

Pioneer.PSM

Pioneer.PrecursorBin

Pioneer.PrecursorBinFragment

Pioneer.addPrecursorID!

Pioneer.countSulfurLoss

Pioneer.countSulfurs!

Pioneer.digest

Pioneer.getA

Pioneer.getAnnotationToID

Pioneer.getB

Pioneer.getFragAbundance!

Pioneer.getFragBounds

Pioneer.getImmoniumToSulfurDict

Pioneer.getIonAnnotationDict

Pioneer.getIonAnnotationSet

Pioneer.getModIndex

Pioneer.getModName

Pioneer.getMzBounds

Pioneer.getNumeric

Pioneer.getPrecursorIsotopeSet

Pioneer.getProtAbundance

Pioneer.getS

Pioneer.make_request

Pioneer.matchPeaks!

Pioneer.multi_request

Pioneer.parseAnnotation

Pioneer.parseAnnotation

Pioneer.parseBatchToTable

Pioneer.parseFasta

Pioneer.parseInternalIon

Pioneer.parseMods

Pioneer.predictFragments

Pioneer.setMatch!

Pioneer.setNearest!

Pioneer.FragmentIndex — Type

FragmentIndex{T<:AbstractFloat}

A fragment index for an MSFragger-style/fragment-centric search. Contains fragment bins that indicate the highest and lowest m/z of a fragment ion in the bin. Each FragBin links to a PrecursorBin. A precursor bin has the precursor m/z's and peptide ID's for each fragment ion.

Fields

fragment_bins::Vector{FragBin{T}} – Lowest m/z of a fragment ion in the FragBin

precursor_bins::Vector{PrecursorBin{T}} – Highest m/z of a fragment ion in the FragBin

Examples

FragmentIndex(T::DataType, M::Int, N::Int) = FragmentIndex(fill(FragBin(), N), fill(PrecursorBin(T, M), N))

GetterMethods

getFragmentBin(fi::FragmentIndex, bin::Int) = fi.fragment_bins[bin]

getPrecursorBin(fi::FragmentIndex, bin::Int64) = fi.precursor_bins[bin]

Methods

setFragmentBin!(fi::FragmentIndex, bin::Int64, frag_bin::FragBin)

setPrecursorBinItem!(fi::FragmentIndex{T}, bin::Int64, index::Int64, precbinitem::PrecursorBinItem{T}) where {T<:AbstractFloat}

Pioneer.FragmentMatch — Type

FragmentMatch

Type that represents a match between a fragment ion and a mass spectrum peak

Fields

transition::Transition – Represents a fragment ion

intensity::Float32 – Intensity of the matching peak

match_mz::Float32 – M/Z of the matching empirical peak. NOT the transition mass and may differ from getMZ(transition) by some error

count::UInt8 – Number of matches (may want to count the number of matches if the spectrum is researched at different mass offsets, such as in a cross correlation score)

peak_int::Int64 – Index of the matching peak in the mass spectrum.

Examples

`FragmentMatch(transition::Transition, intensity::Float32, mass::Float32, count::UInt8, peak_ind::Int64) – default internal constructor

FragmentMatch() – constructor for null/empty precursor

GetterMethods

getMZ(f::FragmentMatch) = getMZ(f.transition)

getLow(f::FragmentMatch) = getLow(f.transition)

getHigh(f::FragmentMatch) = getHigh(f.transition)

getPrecID(f::FragmentMatch) = getPrecID(f.transition)

getCharge(f::FragmentMatch) = getCharge(f.transition)

getIsotope(f::FragmentMatch) = getIsotope(f.transition)

getIonType(f::FragmentMatch) = getIonType(f.transition)

getInd(f::FragmentMatch) = getInd(f.transition)

Pioneer.PSM — Type

PSM

Every concrete instance of type `T` that inherits from `PSM` (peptide spectrum match) should implement the following methods:

ScoreFragmentMatches!(results::UnorderedDictionary{UInt32, T}, matches::Vector{FragmentMatch})).

ModifyFeatures!(score::T, transition::Transition, mass::Union{Missing, Float32}, intensity::Union{Missing, Float32})

makePSMsDict((::T) - Creats the score dictionary

Score!(PSMsdict::Dict, unscoredPSMs::Vector{T}) - Calculates features for each PSM in Vector{T} and adds them to the score dictionary. (See examples)

Examples

PSMdict = makeScoreDict(T) Dict( :FeatureA => [], :FeatureB => [], ... ) Score!(PSMdict, Vector{T<:PSMFeatures}) Dict( :FeatureA => [a, a, ...], :FeatureB => [b, b, ...], ... )

Pioneer.PrecursorBin — Type

PrecursorBin{T<:AbstractFloat}

Container for precursor Items.

Fields

precs::Vector{PrecursorBinItem{T}} – Precursor items

Examples

PrecursorBin(T::DataType, N::Int) = PrecursorBin(Vector{PrecursorBinItem{T}}(undef, N)) – Constructor

GetterMethods

getPrecursors(pb::PrecursorBin) = pb.precs

Methods

setPrecursor!(pb::PrecursorBin, index::Int, pbi::PrecursorBinItem)

Pioneer.PrecursorBinFragment — Type

PrecursorBinItem{T<:AbstractFloat}

Item in a precursor bin. Minimal information required to know if the fragment has a correct precursor mass, and if so, what the precursor ID is.

Fields

prec_id::UInt32 – Unique identifier for the precursor

prec_mz::T – m/z of the precursor

Examples

GetterMethods

getPrecID(pbi::PrecursorBinItem) = pbi.prec_id

getPrecMZ(pbi::PrecursorBinItem) = pbi.prec_mz

Pioneer.addPrecursorID! — Method

addPrecursorID!(df::DataFrame, startprecidx::UInt32, fragsperprec::Int64) adds precursor id's to the koina output formated as a dataframe

Input

-'df::DataFrame' has columns :intensities, :mz, and :annotation -'startprecidx::UInt32' precursoridx of the first precursor -'fragsper_prec::Int64' number of fragments predicted per precursor. Must be the same for all precursors

Output

-'::DataFrame'

Pioneer.countSulfurLoss — Method

countSulfurLoss(molecular_formula::SubString{String},isloss::Bool)

Given a set of strings representing ion annotations, map each one to a unique 16-bit unsigned integer. Return a Julia base Dict mapping the annotations to the integers.

Input

-'molecular_formula::AbstractString' – A molecualr formula like "-2CH3SOH", or "+CH3". Must have sign in front

Output

8-bit integer representing the number of sulfurs added or subtracted from the fragment by the loss -'::Int8'

Examples

julia> countSulfurLoss("+2CH3SOH") 2 julia> countSulfurLoss("-2CH3SOH") -2

Pioneer.countSulfurs! — Method

countSulfurs(plain_sequence::AbstractString, 
                mods_iterator::Base.RegexMatchIterator,
                mods_to_sufur_diff::Dict{String, Int8})::Int8

Counts the number of Sulfurs in the sequence (assumex M/S are the only Sulfur containing characters). Adds or subtracts any sulfurs contributed by the modifications.

Input

-'plainsequence::AbstractString' – Plaine peptide sequence -'modsiterator::Base.RegexMatchIterator' – Iteratres through sequence modifications that might contribute/subtract sulfurs -'modstosufur_diff::Dict{String, Int8}' – maps modification name to the number of sulfurs gained or lost

Output

-'::UInt8' Number of Sulfurs in the precursor (if modstosulfur_diff is incorrectly specified there might be an error because UInt8 doesn't allow for negative values)

Examples

julia> countSulfurs( "LLDYRTIMHDENK", Base.RegexMatchIterator(r"(?<=().*?(?=))", "1(7,M,Oxidation)(13,K,AnExampleMod)", false), Dict("AnExampleMod" => Int8(2)) ) 3

Pioneer.digest — Method

digest(sequence::AbstractString; regex::Regex = r"[KR][^P|$]", max_length::Int = 40, min_length::Int = 8, missed_cleavages::Int = 1)

Given an amino acid sequence, sequence, and a regular expression that matches enzymatic cleavage sites, finds all peptide cleavage products from the amino acid sequence. Can set a minimum and maximum length for cleaved peptides. Gets all peptides with 0:N missed cleavages. Default cleavage site regex is for trypsin. Returns a list of strings that are the peptide cleavage products.

sequence::AbstractString – An amino acid sequence

regex::Regex – Cleavage site regex.

max_length::Int – Exclude peptides longer than this

min_length::Int – Exclude peptides shorter than this

missed_cleavages::Int – Get peptides with 0 to this many missed cleavages

Pioneer.getA — Method

getA(N::Int)

Design matrix for maxLFQ. See the following

* Yu F, Haynes SE, Nesvizhskii AI. IonQuant Enables Accurate and Sensitive Label-Free Quantification With FDR-Controlled Match-Between-Runs. Mol Cell Proteomics. 2021;20:100077. doi: 10.1016/j.mcpro.2021.100077. Epub 2021 Apr 2. PMID: 33813065; PMCID: PMC8131922.
* https://rdrr.io/cran/iq/man/maxLFQ.html
* Cox J, Hein MY, Luber CA, Paron I, Nagaraj N, Mann M. Accurate proteome-wide label-free quantification by delayed normalization and maximal peptide ratio extraction, termed MaxLFQ. Mol Cell Proteomics. 2014 Sep;13(9):2513-26. doi: 10.1074/mcp.M113.031591. Epub 2014 Jun 17. PMID: 24942700; PMCID: PMC4159666.

Pioneer.getAnnotationToID — Method

getAnnotationToID(answer_dict::Dict{String, PioneerFragAnnotation})

Takes a dictionary mapping fragment annotations to PioneerFragAnnotation type. Returns a vector mapping integer values to the PioneerFragAnnotations. Also returns a dictionary mapping the fragment annotations (strings) to their integer keys

Input

-'answer_dict::Dict{String, PioneerFragAnnotation}' – maps fragment annotations to PioneerFragAnnotation

Output

A Set of strings -'::Vector{PioneerFragAnnotation}' -'::Dict{String, UInt16}'

Examples

julia> testdict Dict{String, PioneerFragAnnotation} with 4 entries: ""p-2CH3SOH-H2O"" => PioneerFragAnnotation('p', 0x00, 0x01, 0x00, false, false, -2) ""IEA"" => PioneerFragAnnotation('', 0x00, 0x01, 0x00, false, true, 0) ""y1-NH3"" => PioneerFragAnnotation('y', 0x01, 0x01, 0x00, false, false, 0) ""Int/-CO^2+i"" => PioneerFragAnnotation('_', 0x00, 0x02, 0x01, true, false, 0)

julia> idtoannotation, annotationtoid = getAnnotationToID(test_dict);

julia> idtoannotation 4-element Vector{PioneerFragAnnotation}: PioneerFragAnnotation('p', 0x00, 0x01, 0x00, false, false, -2) PioneerFragAnnotation('', 0x00, 0x01, 0x00, false, true, 0) PioneerFragAnnotation('y', 0x01, 0x01, 0x00, false, false, 0) PioneerFragAnnotation('', 0x00, 0x02, 0x01, true, false, 0)

julia> annotationtoid Dict{String, UInt16} with 4 entries: ""p-2CH3SOH-H2O"" => 0x0001 ""IEA"" => 0x0002 ""y1-NH3"" => 0x0003 ""Int/-CO^2+i"" => 0x0004

Pioneer.getB — Method

getB(S::Matrix{Union{Missing, T}}, N::Int, M::Int) where {T<:Real}

Column vector of sum of median peptide ratios. See the following references.

Yu F, Haynes SE, Nesvizhskii AI. IonQuant Enables Accurate and Sensitive Label-Free Quantification With FDR-Controlled Match-Between-Runs. Mol Cell Proteomics. 2021;20:100077. doi: 10.1016/j.mcpro.2021.100077. Epub 2021 Apr 2. PMID: 33813065; PMCID: PMC8131922.

https://rdrr.io/cran/iq/man/maxLFQ.html

Cox J, Hein MY, Luber CA, Paron I, Nagaraj N, Mann M. Accurate proteome-wide label-free quantification by delayed normalization and maximal peptide ratio extraction, termed MaxLFQ. Mol Cell Proteomics. 2014 Sep;13(9):2513-26. doi: 10.1074/mcp.M113.031591. Epub 2014 Jun 17. PMID: 24942700; PMCID: PMC4159666.

Pioneer.getFragAbundance! — Method

getFragAbundance!(isotopes::Vector{Float64}, iso_splines::IsotopeSplineModel{Float64}, frag::isotope{T, I}, prec::isotope{T, I}, pset::Tuple{I, I}) where {T<:Real,I<:Integer}

Get the relative intensities of fragment isotopes starting at M+0. Fills isotopes in place. isotopes[1] is M+0, isotopes[2] is M+1, etc. Based on Goldfarb et al. 2018 Approximating Isotope Distributions of Biomolecule Fragments CS Omega 2018, 3, 9, 11383-11391 Publication Date:September 19, 2018 https://doi.org/10.1021/acsomega.8b01649

Input

isotopes::Vector{Float64}: – Vector to hold relative abundances of fragment isotopes.

iso_splines::IsotopeSplineModel{Float64} – Splines from Goldfarb et. al. that return isotope probabilities given the number of sulfurs and average mass

frag::isotope{T, I} – The fragment isotope

prec::isotope{T, I} – The precursor isotope

pset::Tuple{I, I} – The first and last precursor isotope that was isolated. (1, 3) would indicate the M+1 through M+3 isotopes were isolated and fragmented.

Output

Fills isotopes in place with the relative abundances of the fragment isotopes. Does not normalize to sum to one!

Notes

See methods from Goldfarb et al. 2018

Algorithm

Examples

Pioneer.getFragBounds — Method

getFragBounds(
    center_mass::AbstractVector{Float32},
    isolation_width::AbstractVector{Float32},

    low_frag_mass::AbstractVector{Float32},
    high_frag_mass::AbstractVector{Float32}
) where {T<:AbstractFloat}

Given the isolation centers and widths of MS2 scans from a raw file and the corresponding scan frange for these MS2 spectra, learns a linear model to predict the maximum and minimum size (m/z) fragment peak given the precursor m/z.

Outputs: FragBoundModel

Essentially a tuple of polynomials. One to predict the high fragment m/z given the precursor m/z

and the other to predict the low fragment m/z

Pioneer.getImmoniumToSulfurDict — Method

getNumeric(immonium_table_path::String)

Given a tab delmited table, maps immonium ion annotations to the respective number of Sulfur's in the ion. See Examples below.

Input

-'immoniumtablepath::String'

Output

-'::Dict{String, String}'

Examples

julia> getImmoniumToFormulaDict("data/immonium.txt") Dict{String, Int8} with 64 entries: "IQA" => 0x0 "IWF" => 0x0 "ILC" => 0x0 "IQB" => 0x0

Pioneer.getIonAnnotationDict — Method

getIonAnnotationDict(ion_annotation_set::Set{String})

Given a set of strings representing ion annotations, map each one to a unique 16-bit unsigned integer. Return a Julia base Dict mapping the annotations to the integers.

Input

-'ionannotationset::String' – [""b9-H2O^2+3i"",...,""y10-CH3SOH-H2O+i""]

Output

A base julia Dict mappign ion annotations to unique 16-bit unsigned integers. -'::Dict{String, UInt32}'

Examples

testset = getIonAnnotationSet(testlib) julia> getIonAnnotationDict(test_set) Dict{String, UInt32} with 4364 entries: ""b20-H2O^2+3i"" => 0x000003c6 ""b13^3+i"" => 0x00000217 ""y5-C2H5NOS"" => 0x00000f85 ""b9-H2O^2+3i"" => 0x00000639 ""b12-CH3SOH-H2O^3"" => 0x000001a5

Pioneer.getIonAnnotationSet — Method

getIonAnnotationSet(lib_path::String)

Extracts unique ion annotations from an .msp (MSPiano) library and returns a set of unique annotations. Special parsing required for internal ions. For example ""Int/PDTAENLPAmY-CO^2+i/7"" poses a problem because there are many unique sequence and internal annotation could have. We just want to extract the loss/isotope/charge information. So we would parse ""Int/PDTAENLPAmY-CO^2+i/7"" => "-CO^2+i".

Input

-'lib_path::String' – "path/to/lib.msp"

Output

A Set of strings -'::Set{String}'

Examples

testset = getIonAnnotationSet(testlib) julia> test_set Set{String} with 4364 elements: ""b9-H2O^2+3i"" ""y17-CH3SOH-H2O^2"" ""y16-CH3SOH^4"" ""y17-H2O-NH3^4+i""

Pioneer.getModIndex — Method

getModIndex(mod_string::AbstractString)::UInt8

Given the a modification string "7,M,Oxidation", parse out the index

Input

-'mod_string::AbstractString' – Modification string

Output

-'::UInt8' modifictaion index

Examples

julia> getModName("7,M,Oxidation") 0x07

Pioneer.getModName — Method

getModName(mod_string::AbstractString)::String

Given the a modification string "7,M,Oxidation", parse out the name

Input

-'mod_string::AbstractString' – Modification string

Output

-'::String' modifictaion name

Examples

julia> getModName("7,M,Oxidation") "Oxidation"

Pioneer.getMzBounds — Method

getMzBounds(mem::MassErrorModel, mass::Float32)

Given a theoretical mass and a MassErrorModel, gets the minimum and maximum bounds for an expected empirical mass. Critically, assumes the empirical mass has already been corrected using mass offset. See getCorrectedMz.

Input

mem::MassErrorModel: – Model for the mass error of an ion

mass::Float32 – A theoretical mass/mz to get boundaries for

Output

Tuple{Float32, Float32} A tuple with the lower and upper boundary respectively.

Notes

Suppose the mass error was 3 ppm. And the tolerance was 10 ppm and 5ppm on the left and right hand sides respectively.

A theoretical mass of 1000000.0f0 m/z, would have a tolerance of (999990.0f0, 1000005.0f0). A theoretical mass falling

Algorithm

Examples

Pioneer.getNumeric — Method

getNumeric(substring::SubString{String}, default::UInt8)

Finds the first numeric subset of a string. Parses to an 8-bit unsigned integer. If no numeric subset is round returns a default value.

Input

-'substring::AbstractString' -'defualt::UInt8'

Output

-'::UInt8'

Examples

julia> getNumeric("x101x", zero(UInt8)) 0x65 julia> getNumeric("x101x102", zero(UInt8)) 0x65 julia> getNumeric("xxxxx", zero(UInt8)) 0x00 julia> getNumeric("xxxxx", one(UInt8)) 0x01

Pioneer.getPrecursorIsotopeSet — Method

getPrecursorIsotopeSet(prec_mz::T, prec_charge::U, window::Tuple{T, T})where {T<:Real,U<:Unsigned}

Given the quadrupole isolation window and the precursor mass and charge, calculates which precursor isotopes were isolated

Input

prec_mz::T: – Precursor mass-to-charge ratio

prec_charge::U – Precursor charge state

window::Tuple{T, T} – The lower and upper m/z bounds of the quadrupole isolation window

Output

A Tuple of two integers. (1, 3) would indicate the M+1 through M+3 isotopes were isolated and fragmented.

Notes

See methods from Goldfarb et al. 2018

Algorithm

Examples

Pioneer.getProtAbundance — Method

getProtAbundance(protein::String, peptides::AbstractVector{String}, experiments::AbstractVector{UInt32}, abundance::AbstractVector{Union{T, Missing}},
                        protein_out::Vector{String}, peptides_out::Vector{String}, experiments_out::Vector{UInt32}, log2_abundance_out::Vector{Float64}) where {T <: Real}

Estimates protein-level abundances from peptide level abundances using the MaxLFQ algorithm

Yu F, Haynes SE, Nesvizhskii AI. IonQuant Enables Accurate and Sensitive Label-Free Quantification With FDR-Controlled Match-Between-Runs. Mol Cell Proteomics. 2021;20:100077. doi: 10.1016/j.mcpro.2021.100077. Epub 2021 Apr 2. PMID: 33813065; PMCID: PMC8131922.

https://rdrr.io/cran/iq/man/maxLFQ.html

Cox J, Hein MY, Luber CA, Paron I, Nagaraj N, Mann M. Accurate proteome-wide label-free quantification by delayed normalization and maximal peptide ratio extraction, termed MaxLFQ. Mol Cell Proteomics. 2014 Sep;13(9):2513-26. doi: 10.1074/mcp.M113.031591. Epub 2014 Jun 17. PMID: 24942700; PMCID: PMC4159666.

Input

peptides, experiments, and abundance each have length N

protein::String – Protien Name

peptides::AbstractVector{String} – Peptide names. Same length as experiments and abundance

experiments::AbstractVector{UInt32} – MS file id's. Same length as peptides and abundance

abundance::AbstractVector{Union{T, Missing}} – Abundances of the peptides. Same lenght as peptides and experiments

protein_out::Vector{String} – The protein name. Repeated N times for each experiment

peptides_out::Vector{String} – List of lists of detected peptides for each of the N experiments. [PEPA;PEPB PEPA;PEPC ...]

experiments_out::Vector{UInt32} – List of expeiments

log2_abundance_out::Vector{Float64} – List of log2 estimated protein abundances for each experiment.

Examples.

julia> using DataFrames

julia> prot = DataFrame(Dict( :peptides => ["A","A","A","B","B","B","C","C","C","D","D","D"], :protein => append!(split(repeat("A",9), ""), ["B","B","B"]), :file_idx => UInt32[1, 2, 3, 1, 2, 3, 1, 2, 3, 1, 2, 3], :abundance => [10, 20, 40, 1, 2, 4, 100, 200, missing, 1000, 2000, 3000], ))

12x4 DataFrame Row │ abundance file_idx peptides protein │ Int64? UInt32 String SubStrin… ─────┼────────────────────────────────────────── 1 │ 10 1 A A 2 │ 20 2 A A 3 │ 40 3 A A 4 │ 1 1 B A 5 │ 2 2 B A 6 │ 4 3 B A 7 │ 100 1 C A 8 │ 200 2 C A 9 │ missing 3 C A 10 │ 1000 1 D B 11 │ 2000 2 D B 12 │ 3000 3 D B

julia> function testLFQ(prot) out = Dict( :protein => String[], :peptides => String[], :log2_abundance => Float64[], :experiments => UInt32[], )

       for (protein, data) in pairs(groupby(prot, :protein))
           getProtAbundance(string(protein[:protein]), 
                               collect(data[!,:peptides]), 
                               collect(data[!,:file_idx]), 
                               collect(data[!,:abundance]),
                               out[:protein],
                               out[:peptides],
                               out[:experiments],
                               out[:log2_abundance]
                           )
       end
       out
   end

testLFQ (generic function with 1 method)

julia> out = testLFQ(prot) Dict{Symbol, Vector} with 4 entries: :protein => ["A", "A", "A", "B", "B", "B"] :peptides => ["A;B;C", "A;B;C", "A;B", "D", "D", "D"] :log2_abundance => [3.15526, 4.15526, 5.15526, 9.96578, 10.9658, 11.5507] :experiments => UInt32[0x00000001, 0x00000002, 0x00000003, 0x00000001, 0x00000002, 0x00000003]

Pioneer.getS — Method

getS(peptides::AbstractVector{String}, peptides_dict::Dict{String, Int64}, experiments::AbstractVector{UInt32}, experiments_dict::Dict{UInt32, Int64}, abundance::AbstractVector{Union{T, Missing}}, M::Int, N::Int) where {T<:Real}

Get MxN matrix of intensities of each peptide in each experiment. Rows stand for experiments and columns for peptides. If the j'th peptide was not seen in the i'th experiment, then S[i, j] == missing.

Input

peptides::AbstractVector{String} – MxN List of peptides for the protein.

peptides_dict::Dict{String, Int64} – Maps N peptides by name to column numbers in S

experiments::AbstractVector{UInt32} – MxN List of experiments

experiments_dict::Dict{UInt32, Int64 – Maps M experiment IDs by name to rows numbers in S

abundance::AbstractVector{Union{T, Missing}} – MxN list of peptide abundances

M::Int – Number of experiments/rows in S

N::Int – Number of peptides/columns in S

Examples

Pioneer.make_request — Method

make_request(json_data::String,model_url::String;max_attempts::Int64 = 100)

Send a request to the koina server to predict fragment intensities

Input

-'jsondata::String' json formatted input with peptide sequences, charges, collision energies and instrument types -'modelurl::String' url for a koina model. See https://koina.wilhelmlab.org/docs -'max_attempts::Int64' maximum number of times to attempt the request before abandoning

Output

-'::Vector{Dict{String,Any}}'

Examples

From the Koina Website

curl "https://koina.wilhelmlab.org:443/v2/models/UniSpec/infer" -d '{ "id": "0", "inputs": [ {"name": "peptidesequences", "shape": [2,1], "datatype": "BYTES", "data": ["YYHTLFTHSLPK", "AGC[UNIMOD:4]FS[UNIMOD:27]PK"]}, {"name": "precursorcharges", "shape": [2,1], "datatype": "INT32", "data": [1,2]}, {"name": "collisionenergies", "shape": [2,1], "datatype": "FP32", "data": [25, 25]}, {"name": "instrumenttypes", "shape": [2,1], "datatype": "BYTES", "data": ["LUMOS", "QE"]}] }'

To format and input in Julia, do the following

julia> databatch = Dict( "id" => "0", "inputs" => [ Dict("name" => "peptidesequences", "shape" => [nrows,1], "datatype" => "BYTES", "data" => getModifiedSequence.(peptidetable[!,:sequence], "", peptidetable[!,:mods])), Dict("name" => "precursorcharges", "shape" => [nrows,1], "datatype" => "INT32", "data" => peptidetable[!,:precursorcharge]), Dict("name" => "collisionenergies", "shape" => [nrows,1], "datatype" => "FP32", "data" => peptidetable[!,:collisionenergy]), Dict("name" => "instrumenttypes", "shape" => [nrows,1], "datatype" => "BYTES", "data" => ["QE" for _ in range(1, nrows)]) ] ) julia> jsondata = JSON.json(databatch) julia> modelurl = "=>"https://koina.wilhelmlab.org:443/v2/models/UniSpec/infer" julia> predictedfragmentsjson = makerequest(jsondata, model_url)

Pioneer.matchPeaks! — Method

function matchPeaks!(matches::Vector{FragmentMatch}, Transitions::Vector{Transition}, masses::Vector{Union{Missing, Float32}}, intensities::Vector{Union{Missing, Float32}}, δ::Float64)

Finds the best matching peak in a mass spectrum for each transition/fragment ion supplied if the match is within the fragment tolerance. Adds each FragmentMatch to matches if not already present. Otherwise, modifies an existing match (see setFragmentMatch!).

Input

matches::Vector{FragmentMatch}: – A list representing fragment ions that match peaks in the mass spectrum (masses)

Transitions::Vector{Transition – A list of fragment ions to search for in the spectrum (masses). MUST BE SORTED IN ASCENDING ORDER BY getMZ(transition)

masses::Vector{Union{Missing, Float32}} – Mass list from a centroided mass spectrum. MUST BE SORTED IN ASCENDING ORDER.

intensities::Vector{Union{Missing, Float32}} – The intensity list from a centroided mass spectrum. Must be the same length as masses

δ::Float64 – A mass offset that can be applied to each mass in masses

Output

Modifies matches[match] if match is <= lenth(matches). Otherwise adds a new FragmentMatch at matches[match]

Notes

Updating a match in matches could be useful if researching the same spectra many times at different mass offsets with matchPeaks! This could be done to calculate a cross correlatoin score for example. If a spectrum is only searched once, then matches should only be added to matches and existing ones never modified.

The fragment tolerance is specified by each Transition. A Transition<:Ion has a field mz::MzFeature which specifies the monoisotopic mass and also upper and lower bounds (the tolerance). See getLow(ion::Ion) and getHigh(ion::Ion). This is why the user need not supply a fragment tolerance to matchPeaks!

masses and intensities contain type unions Union{Missing, Float32}. This method does nothing to check for Missing values, and indeed, it is assumed that there are none, and the presence of any Missing values will cause an error. The reason for the type union is an idiosyncracy of the Arrow.jl package and how it implements nested data types in Arrow files.

Algorithm

Given a list of fragment ions and a centroided mass spectrum both sorted by m/z, it is efficient to search the spetrum for matches in a "single pass" through the spectrum. If there are T transitions and P peaks should be O(T+P). If there are multiple peaks within the tolerance for a given fragment ion, the peak closest in m/z to the fragment ion is chosen. It is possible to assign the same peak to multiple fragment ions, but each fragment ion is only assigned to 0 or 1 peaks.

Examples

Pioneer.multi_request — Method

multirequest(jsondatalist::Vector{String}, modelurl::String; max_attempts::Int64 = 100)

Asyncrhonously make a series of requests to the koina server

Input

-'jsondatalst::Vector{String}' list of json formatted inputs with peptide sequences, charges, collision energies and instrument types -'modelurl::String' url for a koina model. See https://koina.wilhelmlab.org/docs -'maxattempts::Int64' maximum number of times to attempt the request before abandoning

Output

-'::Vector{Vector{Dict{String,Any}}}'

Pioneer.parseAnnotation — Method

parseAnnotation(annotation::String, immonium_to_sulfur)

Extracts unique ion annotations from an .msp (MSPiano) library and returns a set of unique annotations. Special parsing required for internal ions. For example ""Int/PDTAENLPAmY-CO^2+i/7"" poses a problem because there are many unique sequence and internal annotation could have. We just want to extract the loss/isotope/charge information. So we would parse ""Int/PDTAENLPAmY-CO^2+i/7"" => ""Int/-CO^2+i"".

Input

-'lib_path::String' – "path/to/lib.msp"

Output

A Set of strings -'::Set{String}'

Examples

testset = getIonAnnotationSet(testlib) julia> test_set Set{String} with 4364 elements: ""b9-H2O^2+3i"" ""y17-CH3SOH-H2O^2"" ""y16-CH3SOH^4"" ""y17-H2O-NH3^4+i""

Pioneer.parseAnnotation — Method

parseAnnotation(annotation_pieces::Vector{RegexMatch},internal::Bool,immonium::Bool)

Extracts unique ion annotations from an .msp (MSPiano) library and returns a set of unique annotations. Special parsing required for internal ions. For example ""Int/PDTAENLPAmY-CO^2+i/7"" poses a problem because there are many unique sequence and internal annotation could have. We just want to extract the loss/isotope/charge information. So we would parse ""Int/PDTAENLPAmY-CO^2+i/7"" => "-CO^2+i".

Input

-'lib_path::String' – "path/to/lib.msp"

Output

A Set of strings -'::Set{String}'

Examples

testset = getIonAnnotationSet(testlib) julia> test_set Set{String} with 4364 elements: ""b9-H2O^2+3i"" ""y17-CH3SOH-H2O^2"" ""y16-CH3SOH^4"" ""y17-H2O-NH3^4+i""

Pioneer.parseBatchToTable — Method

parseBatchToTable(json_out::Vector{Any})

Parse text output from koina into a dataframe

Input

-'json_out::Vector{Any}' list of dictionaries with the predicted fragment intensities, masses, and names

Output

-'::DataFrame'

Pioneer.parseFasta — Function

    parseFasta(fasta_path::String, parse_identifier::Function = x -> x)::Vector{FastaEntry}

`    Given a fasta file, loads the identifier, description, and sequence into an in memory representation 
`FastaEntry` for each entry. Returns Vector{FastaEntry}

- `fasta_path::String` -- Path to a ".fasta" or ".fasta.gz"
- `parse_identifier::Function ` -- Function takes a String argument and returns a String 
`

Pioneer.parseInternalIon — Method

parseInternalIon(frag_name::String)

Parse out the non-sequence information from the internal ion annotation. See Examples

Input

-'frag_name::String'

Output

-'::SubString{String}'

Examples

julia> parseInternalIon(""Int/PDTAENLPAmY-CO/7"") "-CO" julia> parseInternalIon(""Int/PDTAENLPAmY-CO^2+i/7"") "-CO^2+i"

Pioneer.parseMods — Method

parseMods(mods_string::AbstractString)::Base.RegexMatchIterator

Given the meta data line (starting in "Comment") for a preursor, parse out the fields.

Input

-'mods_string::AbstractString' – String with modifications for hte precursor "1(7,M,Oxidation)(13,K,AnExampleMod)"

Output

-'::Base.RegexMatchIterator}' iterator of RegexMatch for each modification parsed out of the string

Examples

julia> collect( parseMods("1(7,M,Oxidation)(13,K,AnExampleMod)") ) 2-element Vector{RegexMatch}: RegexMatch("7,M,Oxidation") RegexMatch("13,K,AnExampleMod")

Pioneer.predictFragments — Function

predictFragments( peptidetablepath::String, fragsoutpath::String, maxkoinabatches::Int, batchsize::Int, modelname::String; intensity_threshold::Float32 = 0.001f0)

Asyncrhonously make a series of requests to the koina server

Input

-'peptidetablepath::String' path to peptide table in .arrow format. At minimum needs columns :sequence, :precursorcharge, and :collisionenergy -'fragsoutpath::String' path to write the table of predicted fragment ions -'maxattempts::Int64' maximum number of requests to make to koina at once (1-24 are reasonable choices) -'batchsize::Int64' number of precursors to include in a batch (should not exceed what can easily be stored in memory) -'model_name::String' name of the model to use.

Output

-writes the predicted fragment ions to a .arrow table.

julia> test_frags
Arrow.Table with 10217000 rows, 4 columns, and schema:
 :intensities    Float32
 :mz             Float32
 :annotation     String
 :precursor_idx  UInt32

julia> last(DataFrame(test_frags), 5)
5×4 DataFrame
 Row │ intensities  mz        annotation  precursor_idx 
     │ Float32      Float32   String      UInt32        
─────┼──────────────────────────────────────────────────
   1 │  0.00160101  1587.27   y28-NH3^2          165600
   2 │  0.001587    1651.3    y29-NH3^2          165600
   3 │  0.00140379  1354.67   y24-H2O^2          165600
   4 │  0.00136275   966.155  y26^3              165600
   5 │  0.0012414   1009.84   y27^3              165600

Example

Pioneer.setMatch! — Method

setFragmentMatch!(matches::Vector{FragmentMatch}, match::Int, transition::Transition, mass::Float32, intensity::Float32, peak_ind::Int64)

Adds a FragmentMatch to matches if match is not an index in matches, otherwise, updates the match.

Input

hits::Vector{FragmentMatch}: – Represents a fragment ion

match::Int – Index of the match. Must be <=N+1 where N is length(hits)

mass::Float32 – m/z of the emperical peak matched to the transition

intensity::Float32 – intensity of the emperical peak matched to the transition

peak_ind – unique index of the emperical peak matched to the transition

Output

Modifies matches[match] if match is <= lenth(matches). Otherwise adds a new FragmentMatch at matches[match]

Notes

Updating a match in matches could be useful if researching the same spectra many times at different mass offsets with matchPeaks! This could be done to calculate a cross correlatoin score for example. If a spectrum is only searched once, then matches should only be added to matches and existing ones never modified.

Recording the peak_ind could be useful, for example, "chimeric" scoring of a spectrum from a Vector{FragmentMatch} type. The best scoring precursor would have fragments matching to known peak_ind. The Vector{FragmentMatch} could be rescored but excluding FragmentMatches corresponding to those peak_ind's. This would enable a simple chimeric spectra scoring that would not involve completely researching the spectrum (making an additional call to matchPeaks).

Algorithm

Examples

Pioneer.setNearest! — Method

getNearest(transition::Transition, masses::Vector{Union{Missing, Float32}}, peak::Int; δ = 0.01)

Finds the peak (index of masses) nearest in mass to the transition but still within the tolerance (getLow(transition)<masses[peak]<getHigh(transition)). Starts searching at initially supplied peak and increases the index until outside the tolerance. There could be multiple peaks within the tolerance, and this function selects the one with the lowest mass error to the fragment ion.

Input

transition::Transition: – Represents a fragment ion

masses::Vector{Union{Missing, Float32}} – Mass list from a centroided mass spectrum. MUST BE SORTED IN ASCENDING ORDER.

peak::Int – An index for a mass in masses

δ

– A mass offset that can be applied to each mass in masses

Output

An Int representing the index of the m/z in masses nearest in m/z to that of the fragment m/z.

Notes

It is assumed that masses is sorted in ascending order.

In practice, when called from matchPeaks!, masses[peak] will already be within the fragment m/z tolerance.

Usually there will not be another peak in masses where this is true, but it is possible for multiple peaks to fall within the tolerance. The purpose of this function is to select the best peak (closes in mass) when this happens.

Algorithm

Examples